backend/trading/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
amqp==5.3.1
anyio==4.7.0
asgiref==3.8.1
astunparse==1.6.3
attrs==24.2.0
beautifulsoup4==4.12.3
billiard==4.2.1
binance-connector==3.10.0
bs4==0.0.2
CacheControl==0.14.1
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.0
cssselect==1.2.0
cssutils==2.11.1
dateparser==1.2.0
defusedxml==0.8.0rc2
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
firebase-admin==6.6.0
flatbuffers==24.3.25
frozenlist==1.5.0
gast==0.6.0
google-api-core==2.24.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
keras==3.7.0
kombu==5.4.2
libclang==18.1.1
lightgbm==4.5.0
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
more-itertools==10.5.0
msgpack==1.1.0
multidict==6.1.0
namex==0.0.8
newsapi-python==0.2.7
numpy==2.0.2
oauthlib==3.2.2
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pandas==2.2.3
pdfkit==1.0.0
plotly==5.24.1
praw==7.8.1
prawcore==2.4.0
premailer==3.10.0
prompt_toolkit==3.0.48
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybit==5.8.0
pycparser==2.22
pycryptodome==3.21.0
Pygments==2.18.0
PyJWT==2.10.1
pyparsing==3.2.0
python-binance==1.0.25
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.9
python3-openid==3.2.0
pytz==2024.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==1.3.1
rich==13.9.4
rsa==4.9
scikit-learn==1.6.0
scipy==1.14.1
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
soupsieve==2.6
sqlparse==0.5.2
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
threadpoolctl==3.5.0
tweepy==4.14.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
update-checker==0.18.0
uritemplate==4.1.1
urllib3==2.2.3
vaderSentiment==3.3.2
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.1.3
wrapt==1.17.0
xgboost==2.1.3
yagmail==0.15.293
yarl==1.18.3
